DAT-62683: add Kamon + Pr fixes

+++ timbermill-local/src/main/java/com/datorama/oss/timbermill/common/disk/DiskHandler.java
@@ -27,7 +27,7 @@ public interface DiskHandler {
 		return diskHandlerParams;
 	}
 
-	int failedBulksAmount();
+	long failedBulksAmount();
 
 	void close();
 }
